requirements.txt
certifi==2018.8.24
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
future==0.16.0
gunicorn==20.0.0
httplib2==0.11.3
lockfile==0.12.2
mock==3.0.5
oauthlib==2.1.0
olefile==0.46
pandas==0.25.2
paramiko==2.6.0
pexpect==4.6.0
Pillow==6.2.0
protobuf==3.6.1
psycopg2==2.8.4
PyJWT==1.7.0
PyYAML==5.1.2
requests==2.21.0
requests-unixsocket==0.1features = ['bathrooms', 'bedrooms', 'num_features', 'num_photos', 'latitude', 'longitude', 'num_description_words', 'price', 'manager_listings_cnt', 'price_per_room'].5
simplejson==3.16.0
sqlparse==0.3.0
urllib3==1.24.1
